+++ packages/web3-core-method/src/methods/transaction/SendRawTransactionMethod.js
@@ -27,7 +27,7 @@ export default class SendRawTransactionMethod extends AbstractObservedTransactio
      * @param {Utils} utils
      * @param {Object} formatters
      * @param {AbstractWeb3Module} moduleInstance
-     * @param {TransactionObserver} transactionObserver
+     * @param {AbstractTransactionObserver} transactionObserver
      *
      * @constructor
      */