+++ packages/web3-core-method/tests/lib/factories/AbstractMethodFactoryTest.js
@@ -4,7 +4,7 @@ import AbstractMethod from '../../../lib/methods/AbstractMethod';
 import AbstractMethodFactory from '../../../lib/factories/AbstractMethodFactory';
 import AbstractObservedTransactionMethod from '../../../lib/methods/transaction/AbstractObservedTransactionMethod';
 import EthSendTransactionMethod from '../../../src/methods/transaction/EthSendTransactionMethod';
-import TransactionObserver from '../../../src/observers/TransactionObserver';
+import AbstractTransactionObserver from '../../../lib/observers/AbstractTransactionObserver';
 import GetTransactionReceiptMethod from '../../../src/methods/transaction/GetTransactionReceiptMethod';
 import GetBlockByNumberMethod from '../../../src/methods/block/GetBlockByNumberMethod';
 import ChainIdMethod from '../../../src/methods/network/ChainIdMethod';
@@ -16,7 +16,7 @@ jest.mock('web3-core-subscriptions');
 jest.mock('../../../lib/methods/AbstractMethod');
 jest.mock('../../../src/methods/block/GetBlockByNumberMethod');
 jest.mock('../../../src/methods/transaction/GetTransactionReceiptMethod');
-jest.mock('../../../src/observers/TransactionObserver');
+jest.mock('../../../lib/observers/AbstractTransactionObserver');
 jest.mock('../../../src/methods/network/ChainIdMethod');
 jest.mock('../../../src/methods/account/GetTransactionCountMethod');
 
@@ -96,16 +96,14 @@ describe('AbstractMethodFactoryTest', () => {
 
         expect(GetBlockByNumberMethod).toHaveBeenCalledTimes(1);
 
-        expect(NewHeadsSubscription).toHaveBeenCalledTimes(1);
-
-        expect(TransactionObserver).toHaveBeenCalledTimes(1);
+        expect(AbstractTransactionObserver).toHaveBeenCalledTimes(1);
     });
 
     it('calls createMethod and returns a object of type EthSendTransactionMethod', () => {
         new AbstractWeb3Module();
         const moduleInstanceMock = AbstractWeb3Module.mock.instances[0];
         moduleInstanceMock.currentProvider = {supportsSubscriptions: jest.fn()};
-        moduleInstanceMock.currentProvider.supportsSubscriptions.mockReturnValueOnce(false);
+        moduleInstanceMock.currentProvider.supportsSubscriptions.mockReturnValueOnce(true);
 
         expect(abstractMethodFactory.hasMethod('sendEthObserved')).toEqual(true);
 
@@ -117,11 +115,9 @@ describe('AbstractMethodFactoryTest', () => {
 
         expect(GetTransactionReceiptMethod).toHaveBeenCalledTimes(1);
 
-        expect(GetBlockByNumberMethod).toHaveBeenCalledTimes(1);
-
         expect(NewHeadsSubscription).toHaveBeenCalledTimes(1);
 
-        expect(TransactionObserver).toHaveBeenCalledTimes(1);
+        expect(AbstractTransactionObserver).toHaveBeenCalledTimes(1);
 
         expect(ChainIdMethod).toHaveBeenCalledTimes(1);
 
@@ -146,10 +142,8 @@ describe('AbstractMethodFactoryTest', () => {
 
         expect(AbstractObservedTransactionMethod).toHaveBeenCalledTimes(1);
 
-        expect(GetBlockByNumberMethod).toHaveBeenCalledTimes(1);
-
         expect(NewHeadsSubscription).toHaveBeenCalledTimes(1);
 
-        expect(TransactionObserver).toHaveBeenCalledTimes(1);
+        expect(AbstractTransactionObserver).toHaveBeenCalledTimes(1);
     });
 });