+++ packages/web3-eth-contract/src/methods/ContractDeployMethod.js
@@ -27,7 +27,7 @@ export default class ContractDeployMethod extends EthSendTransactionMethod {
      * @param {Utils} utils
      * @param {Object} formatters
      * @param {AbstractWeb3Module} moduleInstance
-     * @param {TransactionObserver} transactionObserver
+     * @param {AbstractTransactionObserver} transactionObserver
      * @param {ChainIdMethod} chainIdMethod
      * @param {GetTransactionCountMethod} getTransactionCountMethod
      *