+++ packages/web3-eth-contract/src/methods/SendContractMethod.js
@@ -29,7 +29,7 @@ export default class SendContractMethod extends EthSendTransactionMethod {
      * @param {Utils} utils
      * @param {Object} formatters
      * @param {AbstractWeb3Module} moduleInstance
-     * @param {TransactionObserver} transactionObserver
+     * @param {AbstractTransactionObserver} transactionObserver
      * @param {ChainIdMethod} chainIdMethod
      * @param {GetTransactionCountMethod} getTransactionCountMethod
      * @param {AllEventsLogDecoder} allEventsLogDecoder