+++ packages/web3-providers/src/providers/HttpProvider.js
@@ -175,11 +175,17 @@ export default class HttpProvider {
                 }
             };
 
-            request.ontimeout = () => {
+            request.addEventListener('timeout', () => {
                 this.connected = false;
 
                 reject(new Error(`Connection error: Timeout exceeded after ${this.timeout}ms`));
-            };
+            });
+
+            request.addEventListener('error', () => {
+                this.connected = false;
+
+                reject(new Error(`Network error ${JSON.stringify(payload)}`));
+            });
 
             try {
                 request.send(JSON.stringify(payload));