+++ CHANGELOG.md
@@ -90,3 +90,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.0.0-alpha.2]
+
+### Added
+
+- ``automine`` and ``Instant Seal`` support added (#2940)
+- Public API layer added to ``web3-core`` module (#3013)
+- getNetworkType method extended with Görli testnet (#3096)
+
+### Changed
+
+- default value of ``transactionConfirmationBlocks`` changed to ``0`` (#3008)
+- ``TransactionObserver`` splitted up to ``HttpTransactionObserver`` and ``SocketTransactionObserver`` (#3008)
+
+### Fixed
+
+- Error handling of the HttpProvider fixed (#2887)
+- Property handling for the Ens module fixed (#3008)
+- ``blockNumber`` check of the received receipt fixed in transaction observers (#3012)
+- Cases for zero valued fromBlock, toBlock filters added (#3109)
+- Topic property of the LogsOptions type extended with null (#3128)